+++ shotmanager/prefs/features.py
@@ -25,6 +25,7 @@ from bpy.types import Operator
 from shotmanager.config import config
 from shotmanager.overlay_tools.interact_shots_stack import shots_stack_prefs
 from shotmanager.overlay_tools.viewport_camera_hud import camera_hud_prefs
+from shotmanager.utils import utils_ui
 
 
 class UAS_ShotManager_Features(Operator):
@@ -37,8 +38,6 @@ class UAS_ShotManager_Features(Operator):
         return context.window_manager.invoke_props_dialog(self, width=420)
 
     def draw(self, context):
-        props = context.scene.UAS_shot_manager_props
-        prefs = context.preferences.addons["shotmanager"].preferences
         layout = self.layout
         layout.alert = True
         layout.label(text="Any change is effective immediately")
@@ -153,8 +152,8 @@ def draw_features_prefs(mode, layout):
     subrow = col.row()
     subrow.scale_x = 1.5
     icon = config.icons_col["ShotManager_NotesData_32"]
-    notesIcon = "TEXT"
-    notesIcon = "WORDWRAP_OFF"
+    # notesIcon = "TEXT"
+    # notesIcon = "WORDWRAP_OFF"
     subrow.prop(props, "display_notes_in_properties", text="", icon_value=icon.icon_id)
     subrow.label(text="Takes and Shots Notes")
 
@@ -176,7 +175,8 @@ def draw_features_prefs(mode, layout):
 
     ################
     ################
-    # layout.separator(factor=separatorVertTopics)
+    if "ADDON_PREFS" != mode:
+        layout.separator(factor=separatorVertTopics)
     layout.label(text="Shot Manager Panels:")
     box = layout.box()
 
@@ -196,7 +196,7 @@ def draw_features_prefs(mode, layout):
     subrow = col.row()
     subrow.scale_x = 1.5
     icon = config.icons_col["ShotManager_CamGPVisible_32"]
-    subrow.prop(prefs, "display_greasepenciltools_in_properties", text="", icon_value=icon.icon_id)
+    subrow.prop(prefs, "display_25D_greasepencil_panel", text="", icon_value=icon.icon_id)
     subrow.label(text="2.5D Grease Pencil Panel")
 
     ################
@@ -204,8 +204,8 @@ def draw_features_prefs(mode, layout):
     subrow = col.row()
     subrow.scale_x = 1.5
     icon = config.icons_col["ShotManager_Retimer_32"]
-    subrow.prop(prefs, "display_render_in_properties", text="", icon="RENDER_ANIMATION")
-    subrow.label(text="Renderer Panel")
+    subrow.prop(prefs, "display_render_panel", text="", icon="RENDER_ANIMATION")
+    subrow.label(text="Render Panel")
 
     ################################################################
     rightCol = boxSplit.column()
@@ -220,96 +220,96 @@ def draw_features_prefs(mode, layout):
     subrow = col.row()
     subrow.scale_x = 1.5
     icon = config.icons_col["ShotManager_Retimer_32"]
-    subrow.prop(prefs, "display_retimer_in_properties", text="", icon_value=icon.icon_id)
+    subrow.prop(prefs, "display_retimer_panel", text="", icon_value=icon.icon_id)
     subrow.label(text="Retimer Panel")
 
-    if "ADDON_PREFS" == mode:
-        return
-
     ################
     ################
-    layout.separator(factor=separatorVertTopics)
+    if "ADDON_PREFS" != mode:
+        layout.separator(factor=separatorVertTopics)
     layout.label(text="Additional Tools in Editors:")
-    separatorLeftWidgets = 1.0
 
     ###################################
     # Sequence timeline ######
     ###################################
-    box = layout.box()
-    subrow = box.row()
+    if "ADDON_PREFS" != mode:
+        box = layout.box()
+        subrow = box.row()
 
-    panelIcon = "TRIA_DOWN" if prefs.seqTimeline_settings_expanded else "TRIA_RIGHT"
-    subrow.prop(prefs, "seqTimeline_settings_expanded", text="", icon_only=True, icon=panelIcon, emboss=False)
+        panelIcon = "TRIA_DOWN" if prefs.seqTimeline_settings_expanded else "TRIA_RIGHT"
+        subrow.prop(prefs, "seqTimeline_settings_expanded", text="", icon_only=True, icon=panelIcon, emboss=False)
 
-    icon = config.icons_col["ShotManager_Retimer_32"]
-    butsubrow = subrow.row()
-    butsubrow.scale_x = 1.5
-    butsubrow.operator(
-        "uas_shot_manager.toggle_seq_timeline_with_overlay_tools",
-        text="",
-        icon="SEQ_STRIP_DUPLICATE",
-        depress=prefs.toggle_overlays_turnOn_sequenceTimeline,
-    )
-    subrow.label(text="Sequence Timeline (in Viewport)")
+        icon = config.icons_col["ShotManager_Retimer_32"]
+        butsubrow = subrow.row()
+        butsubrow.scale_x = 1.5
+        butsubrow.operator(
+            "uas_shot_manager.toggle_seq_timeline_with_overlay_tools",
+            text="",
+            icon="SEQ_STRIP_DUPLICATE",
+            depress=prefs.toggle_overlays_turnOn_sequenceTimeline,
+        )
+        subrow.label(text="Sequence Timeline (in Viewport)")
 
-    if prefs.seqTimeline_settings_expanded:
-        leftCol = box.column()
+        if prefs.seqTimeline_settings_expanded:
+            leftCol = box.column()
 
-        # empty spacer column
-        row = leftCol.row()
-        col = row.column()
-        col.scale_x = 0.25
-        col.label(text=" ")
-        col = row.column(align=True)
+            # empty spacer column
+            row = leftCol.row()
+            col = row.column()
+            col.scale_x = 0.25
+            col.label(text=" ")
+            col = row.column(align=True)
 
-        col.prop(props, "seqTimeline_displayDisabledShots", text="Display Disabled Shots")
-        col.prop(prefs, "seqTimeline_not_disabled_with_overlays")
+            col.prop(props, "seqTimeline_displayDisabledShots", text="Display Disabled Shots")
+            col.prop(prefs, "seqTimeline_not_disabled_with_overlays")
 
     ###################################
     # Interactive Shots Stack ######
     ###################################
-    box = layout.box()
-    subrow = box.row()
+    if "ADDON_PREFS" != mode:
+        box = layout.box()
+        subrow = box.row()
 
-    panelIcon = "TRIA_DOWN" if prefs.intShStack_settings_expanded else "TRIA_RIGHT"
-    subrow.prop(prefs, "intShStack_settings_expanded", text="", icon_only=True, icon=panelIcon, emboss=False)
+        panelIcon = "TRIA_DOWN" if prefs.intShStack_settings_expanded else "TRIA_RIGHT"
+        subrow.prop(prefs, "intShStack_settings_expanded", text="", icon_only=True, icon=panelIcon, emboss=False)
 
-    icon = config.icons_col["ShotManager_Retimer_32"]
-    butsubrow = subrow.row()
-    butsubrow.scale_x = 1.5
-    butsubrow.operator(
-        "uas_shot_manager.toggle_shots_stack_with_overlay_tools",
-        text="",
-        icon="NLA_PUSHDOWN",
-        depress=prefs.toggle_overlays_turnOn_interactiveShotsStack,
-    )
-    subrow.label(text="Interaction Shots Stack (in Timeline)")
+        icon = config.icons_col["ShotManager_Retimer_32"]
+        butsubrow = subrow.row()
+        butsubrow.scale_x = 1.5
+        butsubrow.operator(
+            "uas_shot_manager.toggle_shots_stack_with_overlay_tools",
+            text="",
+            icon="NLA_PUSHDOWN",
+            depress=prefs.toggle_overlays_turnOn_interactiveShotsStack,
+        )
+        subrow.label(text="Interaction Shots Stack (in Timeline)")
 
-    if prefs.intShStack_settings_expanded:
-        shots_stack_prefs.draw_settings(bpy.context, box)
+        if prefs.intShStack_settings_expanded:
+            shots_stack_prefs.draw_settings(bpy.context, box)
 
     ###################################
     # Camera HUD ######################
     ###################################
-    box = layout.box()
-    subrow = box.row()
+    if "ADDON_PREFS" != mode:
+        box = layout.box()
+        subrow = box.row()
 
-    panelIcon = "TRIA_DOWN" if prefs.cameraHUD_settings_expanded else "TRIA_RIGHT"
-    subrow.prop(prefs, "cameraHUD_settings_expanded", text="", icon_only=True, icon=panelIcon, emboss=False)
+        panelIcon = "TRIA_DOWN" if prefs.cameraHUD_settings_expanded else "TRIA_RIGHT"
+        subrow.prop(prefs, "cameraHUD_settings_expanded", text="", icon_only=True, icon=panelIcon, emboss=False)
 
-    icon = config.icons_col["ShotManager_Retimer_32"]
-    butsubrow = subrow.row()
-    butsubrow.scale_x = 1.5
-    butsubrow.operator(
-        "uas_shot_manager.camerahud_toggle_display",
-        text="",
-        icon="CAMERA_DATA",
-        depress=props.camera_hud_display_in_viewports or props.camera_hud_display_in_pov,
-    )
-    subrow.label(text="Camera HUD (in Viewport)")
+        icon = config.icons_col["ShotManager_Retimer_32"]
+        butsubrow = subrow.row()
+        butsubrow.scale_x = 1.5
+        butsubrow.operator(
+            "uas_shot_manager.camerahud_toggle_display",
+            text="",
+            icon="CAMERA_DATA",
+            depress=props.camera_hud_display_in_viewports or props.camera_hud_display_in_pov,
+        )
+        subrow.label(text="Camera HUD (in Viewport)")
 
-    if prefs.cameraHUD_settings_expanded:
-        camera_hud_prefs.draw_settings(bpy.context, box)
+        if prefs.cameraHUD_settings_expanded:
+            camera_hud_prefs.draw_settings(bpy.context, box)
 
     ###################################
     # Frame Range #####################
@@ -322,9 +322,19 @@ def draw_features_prefs(mode, layout):
     butsubrow.prop(prefs, "display_frame_range_tool", text="", icon="CENTER_ONLY")
     subrow.label(text="Frame Range Tool (on Timeline Menu)")
 
-    layout.separator(factor=separatorVertTopics)
+    ###################################
+    # Stamp Info ######################
+    ###################################
 
-    layout.separator(factor=2)
+    box = layout.box()
+    subrow = box.row()
+    subrow.separator(factor=3.5)
+    subrow.enabled = utils_ui.drawStampInfoBut(subrow)
+    subrow.label(text="Stamp Info (1n 3D View tab)")
+
+    layout.separator(factor=separatorVertTopics)
+    if "ADDON_PREFS" != mode:
+        layout.separator(factor=2)
 
 
 _classes = (UAS_ShotManager_Features,)